test log debugging.  fixed startfrom, stations,state,and log complete steps

+++ BurnInControl.Shared/ComDefinitions/Station/AcknowledgeType.cs
@@ -6,7 +6,7 @@ using BurnInControl.Shared.ComDefinitions.MessagePacket;
 namespace BurnInControl.Shared.ComDefinitions.Station;
 
 [JsonConverter(typeof(AcknowledgeTypeJsonConverter))]
-public class AcknowledgeType(string name, byte value) : SmartEnum<AcknowledgeType, byte>(name, value), IPacket {
+public class AcknowledgeType(string name, int value) : SmartEnum<AcknowledgeType, int>(name, value), IPacket {
     public static readonly AcknowledgeType TestStartAck = new AcknowledgeType(nameof(TestStartAck), 0);
     public static readonly AcknowledgeType VersionAck = new AcknowledgeType(nameof(VersionAck), 1);
     public static readonly AcknowledgeType IdAck = new AcknowledgeType(nameof(IdAck), 2);
